Implementing visual code swapover
