+++ megamek/src/megamek/client/ui/swing/dialog/imageChooser/CamoChooserTree.java
@@ -15,29 +15,22 @@
  */
 package megamek.client.ui.swing.dialog.imageChooser;
 
-import java.util.Arrays;
-import java.util.Enumeration;
+
 import java.util.Iterator;
-import javax.swing.JTree;
 import javax.swing.tree.DefaultMutableTreeNode;
 import javax.swing.tree.DefaultTreeModel;
-import javax.swing.tree.TreeSelectionModel;
 
 import megamek.client.ui.swing.tileset.MMStaticDirectoryManager;
 import megamek.common.icons.AbstractIcon;
-import megamek.common.icons.Camouflage;
 
-public class CamoChooserTree extends JTree {
+public class CamoChooserTree extends AbstractIconChooserTree {
     private static final long serialVersionUID = -452869897803327464L;
 
     public CamoChooserTree() {
-        super(); 
-
-        // set up the directory tree (left panel) 
-        getSelectionModel().setSelectionMode(TreeSelectionModel.SINGLE_TREE_SELECTION);
+        super();
 
         DefaultMutableTreeNode root = new DefaultMutableTreeNode(AbstractIcon.ROOT_CATEGORY);
-        root.add(new DefaultMutableTreeNode(Camouflage.NO_CAMOUFLAGE));
+        root.add(new DefaultMutableTreeNode(AbstractIcon.DEFAULT_ICON_FILENAME));
         if (MMStaticDirectoryManager.getCamouflage() != null) {
             if (MMStaticDirectoryManager.getCamouflage().getItemNames("").hasNext()) {
                 root.add(new DefaultMutableTreeNode(AbstractIcon.ROOT_CATEGORY));
@@ -53,46 +46,4 @@ public class CamoChooserTree extends JTree {
         }
         setModel(new DefaultTreeModel(root));
     }
-    
-    /**
-     * This recursive method is a hack: DirectoryItems flattens the directory
-     * structure, but it provides useful functionality, so this method will
-     * reconstruct the directory structure for the JTree.
-     *
-     * @param node
-     * @param names
-     */
-    private void addCategoryToTree(DefaultMutableTreeNode node, String... names) {
-        // Shouldn't happen
-        if (names.length == 0) {
-            return;
-        }
-
-        boolean matched = false;
-        for (Enumeration<?> e = node.children(); e.hasMoreElements();) {
-            DefaultMutableTreeNode childNode = (DefaultMutableTreeNode) e.nextElement();
-            String nodeName = (String) childNode.getUserObject();
-            if (nodeName.equals(names[0])) {
-                if (names.length > 1) {
-                    addCategoryToTree(childNode,
-                            Arrays.copyOfRange(names, 1, names.length));
-                    matched = true;
-                } else {
-                    // I guess we're done? This shouldn't happen, as there
-                    // shouldn't be duplicates
-                }
-            }
-        }
-
-        // If we didn't match, lets create nodes for each name
-        if (!matched) {
-            DefaultMutableTreeNode root = node;
-            for (String name : names) {
-                DefaultMutableTreeNode newNode = new DefaultMutableTreeNode(name);
-                root.add(newNode);
-                root = newNode;
-            }
-        }
-    }
-
 }
