+++ megamek/src/megamek/client/ui/swing/dialog/imageChooser/CamouflagePanel.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright (c) 2020 - The MegaMek Team. All Rights Reserved.
+ *
+ * This file is part of MegaMek.
+ *
+ * MegaMek is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * MegaMek is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with MegaMek. If not, see <http://www.gnu.org/licenses/>.
+ */
+package megamek.client.ui.swing.dialog.imageChooser;
+
+import megamek.common.icons.AbstractIcon;
+import megamek.common.icons.Camouflage;
+
+public class CamouflagePanel extends AbstractIconPanel {
+    //region Constructors
+    public CamouflagePanel() {
+        this(new Camouflage());
+    }
+
+    public CamouflagePanel(AbstractIcon icon) {
+        super(icon, "CamouflagePanel");
+    }
+    //endregion Constructors
+}
