+++ megamek/src/megamek/client/ui/swing/dialog/imageChooser/PortraitChooserTree.java
@@ -1,27 +1,36 @@
+/*
+ * Copyright (c) 2020 - The MegaMek Team. All Rights Reserved.
+ *
+ * This file is part of MegaMek.
+ *
+ * MegaMek is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * MegaMek is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with MegaMek. If not, see <http://www.gnu.org/licenses/>.
+ */
 package megamek.client.ui.swing.dialog.imageChooser;
 
-import java.util.Arrays;
-import java.util.Enumeration;
 import java.util.Iterator;
-
-import javax.swing.JTree;
 import javax.swing.tree.DefaultMutableTreeNode;
 import javax.swing.tree.DefaultTreeModel;
-import javax.swing.tree.TreeSelectionModel;
 
 import megamek.client.ui.swing.tileset.MMStaticDirectoryManager;
 import megamek.common.icons.AbstractIcon;
 
-public class PortraitChooserTree extends JTree {
-    
+public class PortraitChooserTree extends AbstractIconChooserTree {
     private static final long serialVersionUID = 1274949831997174959L;
-    
-    public PortraitChooserTree() {
-        super(); 
 
-        // set up the directory tree (left panel) 
-        getSelectionModel().setSelectionMode(TreeSelectionModel.SINGLE_TREE_SELECTION);
-        
+    public PortraitChooserTree() {
+        super();
+
         DefaultMutableTreeNode root = new DefaultMutableTreeNode(AbstractIcon.ROOT_CATEGORY);
         if (MMStaticDirectoryManager.getPortraits() != null) {
             Iterator<String> catNames = MMStaticDirectoryManager.getPortraits().getCategoryNames();
@@ -34,46 +43,5 @@ public class PortraitChooserTree extends JTree {
             }
         }
         setModel(new DefaultTreeModel(root));
-    }
-    
-    /**
-     * This recursive method is a hack: DirectoryItems flattens the directory
-     * structure, but it provides useful functionality, so this method will
-     * reconstruct the directory structure for the JTree.
-     *
-     * @param node
-     * @param names
-     */
-    private void addCategoryToTree(DefaultMutableTreeNode node, String... names) {
-        // Shouldn't happen
-        if (names.length == 0) {
-            return;
-        }
-
-        boolean matched = false;
-        for (Enumeration<?> e = node.children(); e.hasMoreElements();) {
-            DefaultMutableTreeNode childNode = (DefaultMutableTreeNode) e.nextElement();
-            String nodeName = (String) childNode.getUserObject();
-            if (nodeName.equals(names[0])) {
-                if (names.length > 1) {
-                    addCategoryToTree(childNode,
-                            Arrays.copyOfRange(names, 1, names.length));
-                    matched = true;
-                } else {
-                    // I guess we're done? This shouldn't happen, as there
-                    // shouldn't be duplicates
-                }
-            }
-        }
-
-        // If we didn't match, lets create nodes for each name
-        if (!matched) {
-            DefaultMutableTreeNode root = node;
-            for (String name : names) {
-                DefaultMutableTreeNode newNode = new DefaultMutableTreeNode(name);
-                root.add(newNode);
-                root = newNode;
-            }
-        }
     }
 }
